dom/tests/mochitest/bugs/test_bug583225.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=583225
-->
<head>
  <title>Test for Bug 583225</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=583225">Mozilla Bug 583225</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 583225 **/

function clicked()
{
  ok(true, "Event fired on DocumentFragment.");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

var docFragment = document.createDocumentFragment();
docFragment.foo = "foo";
docFragment.onclick = clicked;
var event = document.createEvent('MouseEvent');
event.initMouseEvent('click', true, true, window,  1, 0, 0, 0, 0, false, false,
                     false, false, 0, docFragment);
docFragment.dispatchEvent(event);

</script>
</pre>
</body>
</html>
